+++ design-system/stories/Button/VerticalCardButton.stories.tsx
@@ -1,5 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import VerticalCardButton from '../../ui/Button/VerticalCardButton';
+
 type Story = StoryObj<typeof meta>;
 
 const meta = {
