+++ design-system/stories/textFields/UnderlineTextField.stories.tsx
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import DefaultTextField from '../../ui/textFields/DefaultTextField';
+import UnderlineTextField from '../../ui/textFields/UnderlineTextField';
 
 const meta = {
-  title: 'TextFields/DefaultTextField',
-  component: DefaultTextField,
+  title: 'TextFields/UnderlineTextField',
+  component: UnderlineTextField,
   parameters: {
     layout: 'centered',
   },
@@ -31,7 +31,7 @@ const meta = {
       description: '입력 이벤트',
     },
   },
-} satisfies Meta<typeof DefaultTextField>;
+} satisfies Meta<typeof UnderlineTextField>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
