+++ design-system/ui/Button/IconButton.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 export interface IconButtonProps {
   iconPath: string;
   onClick: () => void;
