+++ design-system/ui/Button/PrimaryButton.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 interface PrimaryButtonProps {
   label: string; // 버튼 텍스트
   onClick: () => void; // 클릭 핸들러
